pyproject.toml
[project]
name = "teslamate-openclaw-connector"
version = "0.1.0"
description = "OpenClaw Skill connector for TeslaMate via Tailscale"
requires-python = ">=3.12"
dependencies = [
    "paho-mqtt>=2.0.0",
    "httpx>=0.27.0",
    "websockets>=12.0",
    "pyyaml>=6.0",
    "pydantic-settings>=2.0",
]

[project.scripts]
teslamate-connector = "teslamate_connector.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/teslamate_connector"]
